requirements.txt
Django>=4.2,<5.3
Pillow>=10.0

Faker>=19.0
